Сделано добавление счетов
- Изменена логика расчета наличия транзакций у связанных сущностей.
- Сделан расчет баланса счетов.

+++ src/main/java/ru/fess38/finance/controller/AccountController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import ru.fess38.finance.dao.AccountDao;
+import ru.fess38.finance.dao.TransactionDao;
 import ru.fess38.finance.model.Account;
 
 import java.util.List;
@@ -17,6 +18,8 @@ import java.util.List;
 public class AccountController {
   @Autowired
   private AccountDao accountDao;
+  @Autowired
+  private TransactionDao transactionDao;
 
 
   @RequestMapping(value = "/account/get", method = RequestMethod.GET)
@@ -46,7 +49,9 @@ public class AccountController {
 
   @RequestMapping(value = "/account/delete", method = RequestMethod.POST)
   public void delete(@RequestBody Account account) {
-    accountDao.delete(account);
+    if (transactionDao.countByAccount(account) == 0) {
+      accountDao.delete(account);
+    }
   }
 
 }
